validation fix

+++ routes/api/contacts.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const createError = require("http-errors");
-
+const mongoose = require("mongoose");
 const { Contact, schemas } = require("../../models/contact");
 
 const router = express.Router();
@@ -22,15 +22,15 @@ router.get("/", async (req, res, next) => {
 router.get("/:id", async (req, res, next) => {
   try {
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new createError(400, "invalid ID");
+    }
     const result = await Contact.findById(id, "-createdAt -updatedAt");
     if (!result) {
       throw new createError(404, "Not found");
     }
     res.json(result);
   } catch (error) {
-    if (error.message.includes("Cast to ObjectId failed")) {
-      error.status = 404;
-    }
     next(error);
   }
 });
@@ -44,9 +44,6 @@ router.post("/", async (req, res, next) => {
     const result = await Contact.create(req.body, "-createdAt -updatedAt");
     res.status(201).json(result);
   } catch (error) {
-    if (error.message.includes("validation failed")) {
-      error.status = 400;
-    }
     next(error);
   }
 });
@@ -54,6 +51,9 @@ router.post("/", async (req, res, next) => {
 router.delete("/:id", async (req, res, next) => {
   try {
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new createError(400, "invalid ID");
+    }
     const result = await Contact.findByIdAndDelete(id);
     if (!result) {
       throw new createError(404, "Not found");
@@ -67,6 +67,9 @@ router.delete("/:id", async (req, res, next) => {
 router.put("/:id", async (req, res, next) => {
   try {
     const { error } = schemas.add.validate(req.body);
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new createError(400, "invalid ID");
+    }
     if (error) {
       throw new createError(400, error.message);
     }
@@ -91,6 +94,9 @@ router.patch("/:id/favorite", async (req, res, next) => {
       throw new createError(400, "missing field favorite");
     }
     const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new createError(400, "invalid ID");
+    }
     const result = await Contact.findByIdAndUpdate(id, req.body, {
       new: true,
       select: "-createdAt -updatedAt",
